pyproject.toml
[project]
name = "quantilan-agent"
version = "0.1.0"
description = "Quantilan Trading Agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.26.0",
    "aiohttp>=3.9.0",
    "anthropic>=0.40.0",
    "ccxt>=4.2.0",
    "certifi>=2024.0.0",
    "fastapi>=0.110.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.27.0",
    "websockets>=12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "gui"]

[dependency-groups]
dev = []

[tool.uv.sources]
# Pin ccxt.pro if available — comment out if using ccxt only
# ccxt = { git = "https://github.com/ccxt/ccxt" }
